Move PermissionLexerException to Bolt\Exception
